+++ internal/server/server_embedded_ui_dev.go
@@ -1,9 +1,8 @@
 //go:build !prod
-// +build !prod
 
 package server
 
 import "embed"
 
 //go:embed web_dev
-var embeddedFiles embed.FS
+var embeddedWebinterfaceFS embed.FS