+++ internal/server/server_embedded_ui_prod.go
@@ -1,9 +1,8 @@
 //go:build prod
-// +build prod
 
 package server
 
 import "embed"
 
 //go:embed web/build/*
-var embeddedFiles embed.FS
+var embeddedWebinterfaceFS embed.FS